+++ src/org/nfrac/comportex/demos/cortical_io_channel.clj
@@ -7,32 +7,32 @@
             [clojure.core.async :as async
              :refer [chan <!! >!! thread close!]]))
 
-(def spec
+(def params
   {:column-dimensions [40 40]
    :ff-init-frac 0.15
    :ff-potential-radius 1.0
    :proximal {:stimulus-threshold 3}
    :spatial-pooling :local-inhibition
    :activation-level 0.015
-   :distal-vs-proximal-weight 1.0
-   })
+   :distal-vs-proximal-weight 1.0})
 
-(def higher-level-spec
-  (merge spec
+
+(def higher-level-params
+  (merge params
          {:column-dimensions [20 20]
           :proximal {:max-segments 5}}))
 
 (defn split-sentences
   [text]
   (->> (str/split (str/trim text) #"[^\w]*[\.\!\?]+[^\w]*")
-       (mapv #(str/split % #"[^\w']+"))
+       (mapv #(str/split % #"[^\w']+"))))
        ;(mapv #(conj % "."))
-       ))
+
 
 (defn n-region-model
   [api-key cache n]
   (core/regions-in-series n core/sensory-region
-                          (list* spec (repeat higher-level-spec))
+                          (list* params (repeat higher-level-params))
                           {:input (cortical-io-encoder api-key cache)}))
 
 (comment
@@ -99,6 +99,4 @@
   (close! preload-c)
   (close! input-c)
   (close! preloader)
-  (close! stepper)
-
-  )
+  (close! stepper))