+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - The parameter specification map now referred to as params instead of spec.
   To avoid confusion with `clojure.spec` where specs define function usage.
   The breaking part of this is renaming key :spec to :params in LayerOfCells.
+- within layer, state key :matching-seg-paths is now :fully-matching-seg-paths.
 
 ### Added
 - minimal Javascript API in new `js` namespace.