+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 ### Changed
+
+## [0.0.15] - 2016-09-05
+### Changed
 - The parameter specification map now referred to as params instead of spec.
   To avoid confusion with `clojure.spec` where specs define function usage.
   The breaking part of this is renaming key :spec to :params in LayerOfCells.
@@ -91,7 +94,8 @@
 ## [0.0.12] - 2015-12-01
 - Before this I didn't keep a change log.
 
-[Unreleased]: https://github.com/htm-community/comportex/compare/v0.0.14...HEAD
+[Unreleased]: https://github.com/htm-community/comportex/compare/v0.0.15...HEAD
+[0.0.15]: https://github.com/htm-community/comportex/compare/v0.0.14...v0.0.15
 [0.0.14]: https://github.com/htm-community/comportex/compare/v0.0.13...v0.0.14
 [0.0.13]: https://github.com/htm-community/comportex/compare/v0.0.12...v0.0.13
 [0.0.12]: https://github.com/htm-community/comportex/compare/v0.0.10...v0.0.12